Login authentication

+++ app/Http/Controllers/Api/V1/AuthController.php
@@ -15,7 +15,6 @@ class AuthController extends BaseController
         $this->validate($request, [
             'username' => 'required|string',
             'password' => 'required|string|min:5',
-            // 'user_agent' => 'required'
         ]);
 
         $credentials = $request->only('username', 'password');
@@ -28,31 +27,38 @@ class AuthController extends BaseController
 
         $user = $authorization->user();
 
-        if(!$this->authenticated($user)) {
-            return $this->response->errorUnauthorized("This user has logged in somewhere else");
+        if($this->authenticated($user)) {
+            return $this->response->errorUnauthorized("User ini telah login di perangkat lain. Coba login kembali.");
         }
+
+        $expired_time = $authorization->toArray()['expired_at'];
 
         $session = Session::create([
             'user_id' => $user->id,
             'token' => $token,
             'user_agent' => $_SERVER['HTTP_USER_AGENT'],
-            'login_at' => date("Y-m-d H:i:s")
+            'login_at' => date("Y-m-d H:i:s"),
+            'expired_at' => $expired_time,
         ]);
 
         return $this->response->item($authorization, new AuthorizationTransformer);
     }
 
     public function authenticated($user) {
-        $login = Session::where('user_id', $user->id)->count();
-        if($login > 0) {
-            $this->logout($user->id);
+        $login = Session::where('user_id', $user->id)->first();
+
+        if($login) {
+            if(strtotime($login->expired_at) < strtotime(date('Y-m-d H:i:s'))) {
+                $this->logout($user->id);
+                return TRUE;
+            }
+            return TRUE;
+        } else {
             return FALSE;
         }
-        return TRUE;
     }
 
     public function logout($user_id) {
-        $user_session = Session::where('user_id', $user_id)->first();
-        $user_session->delete();
+        $user_session = Session::where('user_id', $user_id)->first()->delete();
     }
 }
